refactor/update-project-name -) Changed package name and directory names.
