+++ app/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.mythicalcreaturesllc.breedingandanimaltracking"
+        applicationId "com.mythcreatures.breedingtracker"
         minSdkVersion 19
         targetSdkVersion 28
         versionCode 1
