+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Breeding and Animal Tracking</string>
+    <string name="app_name">Breeding Tracker</string>
 </resources>
